+++ web/src/components/UpstreamAccountCreatePage.api-key.stories.tsx
@@ -1,0 +1,39 @@
+import type { Meta, StoryObj } from '@storybook/react-vite'
+import {
+  AccountPoolStoryRouter,
+  UpstreamAccountCreatePage,
+  upstreamAccountCreateMetaBase,
+} from './UpstreamAccountCreatePage.story-common'
+
+const meta = {
+  ...upstreamAccountCreateMetaBase,
+  title: 'Account Pool/Pages/Upstream Account Create/API Key',
+} satisfies Meta<typeof UpstreamAccountCreatePage>
+
+export default meta
+
+type Story = StoryObj<typeof meta>
+
+export const NameConflict: Story = {
+  render: () => (
+    <AccountPoolStoryRouter
+      initialEntry={{
+        pathname: '/account-pool/upstream-accounts/new',
+        search: '?mode=apiKey',
+        state: {
+          draft: {
+            apiKey: {
+              displayName: ' team key - staging ',
+              groupName: 'staging',
+              note: 'Conflicts with an existing API Key account name.',
+              apiKeyValue: 'sk-storybookduplicate1234',
+              primaryLimit: '120',
+              secondaryLimit: '500',
+              limitUnit: 'requests',
+            },
+          },
+        },
+      }}
+    />
+  ),
+}
